+++ scripts/test_script.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+npm test "tests/selenium.test_cron.js"
+wait
+npm test "tests/gamelogic.test_cron.js"
+wait
+python3 tests/seleniumPythonTest/test_case/slnm_test_check_title.py
